Using navigator do de-duplicate some code
